requirements.txt
Django
djangorestframework
django-filter
drf-nested-routers
drf_yasg
future
psycopg2
django-cors-headers